requirements.txt
tensorflow
keras
flask
Pillow
pydantic
opencv-python
